# api/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Bina Academy API",
    "description": "## نظرة عامة | Overview\n\nمجموعة الـ APIs الخاصة بمنصة **Bina Academy** لطلبة التوجيهي (الفرع العلمي والأدبي - الصف 11 و 12).\n\nAPI collection for **Bina Academy** tutoring platform (Tawjihi students - grades 11/12, scientific/literary branches).\n\n---\n\n### خصائص عامة | General Properties\n\n- جميع الردود JSON (`Content-Type: application/json`)\n- جميع أرقام الهواتف بالصيغة الفلسطينية `^05\\d{8}$` (مثال: `0591234567`)\n- التواريخ بصيغة ISO 8601\n- المصادقة بـ JWT Bearer (توكينات الطلاب 30 يوم، توكينات الإدمن 7 أيام)\n- الحد العام للطلبات: 100 طلب/دقيقة لكل IP (حسب `@nestjs/throttler`)\n\n### تدفّق العمل الموصى به | Recommended Mobile Flows\n\n1. **تسجيل طالب جديد**: POST `/auth/register` (الهاتف، كلمة المرور، الاسم الكامل، الصف، الفرع) → تصفح المواد مباشرة.\n2. **دخول طالب مسجل**: POST `/auth/login` (الهاتف وكلمة المرور) → تصفح المواد.\n3. **مشاهدة فيديو**: GET `/subjects/:id/videos` → GET `/videos/:id/stream` → POST `/videos/:id/mark-viewed`\n4. **إدمن**: POST `/auth/admin/login` → استعمل الـ token في كل endpoints الـ admin\n",
    "version": "1.0.0",
    "contact": {
      "name": "Bina Academy Backend Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Local development"
    },
    {
      "url": "http://10.0.2.2:3000/api",
      "description": "Android emulator (loopback to host)"
    },
    {
...
    }
  ],
...
  "paths": {
...
    "/subjects": {
      "get": {
        "tags": ["Subjects"],
        "summary": "قائمة المواد للطالب الحالي (حسب صفه وفرعه)",
        "description": "يقرأ `grade` و `branch` من ملف الطالب، ثم يعيد المواد المطابقة فقط مع عدد الفيديوهات المنشورة واسم المعلم. إذا لم يكن قد أكمل الملف الشخصي بعد، يرد بـ 404.\n\nReads the student's `grade` and `branch` from profile and returns matching subjects with published video count and teacher name. Returns 404 if profile incomplete.",
        "operationId": "listSubjects",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "قائمة المواد",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/SubjectListItem" }
                },
                "example": [
                  {
                    "id": "sub_math",
                    "name": "الرياضيات",
                    "grade": "GRADE_12",
                    "branch": "SCIENTIFIC",
                    "teacherName": "أ. خالد عمر",
                    "videoCount": 24,
                    "progressPercent": 75,
                    "locked": true
                  },
                  {
                    "id": "sub_phys",
                    "name": "الفيزياء",
                    "grade": "GRADE_12",
                    "branch": "SCIENTIFIC",
                    "teacherName": "أ. سامي حسن",
                    "videoCount": 18,
                    "progressPercent": 40,
                    "locked": false
                  }
                ]
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": {
            "description": "الطالب لم يكمل الملف الشخصي بعد (لا grade/branch)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": {
                  "statusCode": 404,
                  "message": "يرجى إكمال الملف الشخصي أولاً"
                }
              }
            }
          }
        }
      }
    },
    "/subjects/my": {
      "get": {
        "tags": ["Subjects"],
        "summary": "قائمة موادي المشترك بها الطالب (المكتبة) | Purchased Student Subjects",
        "description": "يعود بالمواد التي اشترك بها الطالب واشتراكه فعال فيها حالياً. مفيد جداً لشاشة 'موادي' في التطبيق لزيادة سرعة الأداء وتقليل استهلاك البيانات.\n\nReturns subjects that the student has purchased (has an active paid subscription for).",
        "operationId": "listPurchasedSubjects",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "قائمة المواد المشترك بها بنجاح",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/SubjectListItem" }
                },
                "example": [
                  {
                    "id": "sub_phys",
                    "name": "الفيزياء",
                    "grade": "GRADE_12",
                    "branch": "SCIENTIFIC",
                    "teacherName": "أ. سامي حسن",
                    "videoCount": 18,
                    "progressPercent": 40,
                    "locked": false
                  }
                ]
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/subjects/{id}/videos": {
      "get": {
        "tags": ["Subjects"],
        "summary": "فيديوهات مادة معيّنة مع حالة القفل",
        "description": "يعرض كل الفيديوهات **المنشورة** للمادة مرتّبة حسب (unitNumber, orderInUnit). كل فيديو يحمل `locked: boolean` — يعتمد على `quota = plan.videosPerSubject` و **الترتيب** (`locked = (index >= quota)`)، وليس على مشاهدات سابقة.\n\n> ⚠️ **قاعدة الحصة**: لو الخطة تسمح بـ 10 فيديوهات، فالأول 10 فيديوهات فقط مفتوحة والباقي `locked: true` — حتى لو لم تُشاهد بعد.\n\nLists PUBLISHED videos in display order (unitNumber, orderInUnit). Each video has `locked: boolean` based on quota (top N unlocked). Locking is determined by **position in the sorted list**, not view history.",
        "operationId": "getSubjectVideos",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "معرّف المادة",
            "schema": { "type": "string", "example": "sub_math" }
          }
        ],
        "responses": {
          "200": {
            "description": "المادة وفيديوهاتها وحصتها",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SubjectVideosResponse" },
                "example": {
                  "subject": { "id": "sub_math", "name": "الرياضيات" },
                  "quota": 10,
                  "videos": [
                    {
                      "id": "vid_001",
                      "title": "النهايات - مقدمة",
                      "description": "تمهيد لمفهوم النهاية",
                      "durationSec": 645,
                      "unitNumber": 1,
                      "orderInUnit": 1,
                      "locked": false
                    },
                    {
                      "id": "vid_002",
                      "title": "النهايات - أمثلة",
                      "durationSec": 720,
                      "unitNumber": 1,
                      "orderInUnit": 2,
                      "locked": false
                    },
                    {
                      "id": "vid_011",
                      "title": "التفاضل - الوحدة 2",
                      "durationSec": 900,
                      "unitNumber": 2,
                      "orderInUnit": 1,
                      "locked": true
                    }
                  ]
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": {
            "description": "المادة غير موجودة",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": {
                  "statusCode": 404,
                  "message": "المادة غير موجودة"
                }
              }
            }
          }
        }
      }
    },
...
    "/dashboard": {
      "get": {
        "tags": ["Users"],
        "summary": "بيانات شاشة الرئيسية للموبايل | Mobile Home Screen Dashboard",
        "description": "يعيد بيانات لوحة معلومات الطالب بما في ذلك تقدمه العام، تفاصيل الفيديو الأخير لمتابعة التعلم، أهداف اليوم مع النسبة والرسالة الترحيبية، حالة الاختبار اليومي، والمواد المقترحة لليوم.",
        "operationId": "getDashboard",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "بيانات لوحة المعلومات بنجاح",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "studentName": { "type": "string" },
                    "generalProgress": { "type": "integer" },
                    "continueLearning": {
                      "type": "object",
                      "nullable": true,
                      "properties": {
                        "videoId": { "type": "string" },
                        "videoTitle": { "type": "string" },
                        "subjectName": { "type": "string" },
                        "unitName": { "type": "string" },
                        "lessonText": { "type": "string" },
                        "durationSec": { "type": "integer" },
                        "timeLeftMin": { "type": "integer" },
                        "progressPercent": { "type": "integer" }
                      }
                    },
                    "todayGoals": {
                      "type": "object",
                      "properties": {
                        "completedCount": { "type": "integer" },
                        "totalCount": { "type": "integer" },
                        "percentage": { "type": "integer" },
                        "text": { "type": "string" }
                      }
                    },
                    "dailyQuiz": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "string" },
                        "title": { "type": "string" },
                        "description": { "type": "string" },
                        "isAvailable": { "type": "boolean" },
                        "buttonText": { "type": "string" },
                        "points": { "type": "integer" }
                      }
                    },
                    "suggestedSubjects": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "string" },
                          "subjectName": { "type": "string" },
                          "teacherName": { "type": "string" },
                          "rating": { "type": "number" }
                        }
                      }
                    },
                    "subjectShopping": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "string" },
                          "name": { "type": "string" },
                          "priceIls": { "type": "number" },
                          "isSubscribed": { "type": "boolean" },
                          "isInCart": { "type": "boolean" },
                          "teacher": {
                            "type": "object",
                            "nullable": true,
                            "properties": {
                              "id": { "type": "string" },
                              "name": { "type": "string" },
                              "avatarUrl": { "type": "string", "format": "uri" },
                              "rating": { "type": "number" }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/goals": {
      "get": {
        "tags": ["Users"],
        "summary": "قائمة أهداف الطالب لليوم والأهداف المؤرشفة | List Student Daily & Archived Goals",
        "description": "يعيد قائمة بأهداف اليوم الحالي (مع حالة الإنجاز) وقائمة بالأهداف السابقة المؤرشفة.",
        "operationId": "listGoals",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "قائمة الأهداف",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "today": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/DailyGoal" }
                    },
                    "archived": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/DailyGoal" }
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "post": {
        "tags": ["Users"],
        "summary": "إضافة هدف جديد لليوم | Create New Daily Goal",
        "description": "يضيف هدفاً جديداً لقائمة مهام اليوم الحالي.",
        "operationId": "createGoal",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title"],
                "properties": {
                  "title": { "type": "string", "example": "قراءة فصل من كتاب الكيمياء" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "تم إنشاء الهدف بنجاح",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DailyGoal" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/goals/{id}": {
      "patch": {
        "tags": ["Users"],
        "summary": "تحديث أو تبديل حالة الهدف | Update Goal Status or Title",
        "description": "يسمح بتحديث حالة إنجاز الهدف (معلم كمنجز/غير منجز) أو تعديل عنوانه.",
        "operationId": "updateGoal",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "completed": { "type": "boolean" },
                  "title": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "تم تحديث الهدف بنجاح",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DailyGoal" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "description": "غير مصرح لك بتعديل هذا الهدف" },
          "404": { "description": "الهدف غير موجود" }
        }
      },
      "delete": {
        "tags": ["Users"],
        "summary": "حذف هدف | Delete Goal",
        "description": "يحذف الهدف المحدد.",
        "operationId": "deleteGoal",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "تم حذف الهدف بنجاح",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "description": "غير مصرح لك بحذف هذا الهدف" },
          "404": { "description": "الهدف غير موجود" }
        }
      }
    },
...
    "/admin/teachers/dashboard": {
      "get": {
        "tags": ["Admin"],
        "summary": "إحصائيات وطلبات المعلمين | Teachers Dashboard Stats & Jobs",
        "description": "يعود بإحصائيات المعلمين، طلبات التوظيف الجديدة وقائمة المعلمين المتميزين. يتطلب توكين ADMIN.\n\nReturns teachers KPI stats, new job applications, and top teachers list. Requires ADMIN JWT.",
        "operationId": "getTeachersDashboard",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "بيانات لوحة المعلمين بنجاح",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "stats": {
                      "type": "object",
                      "properties": {
                        "totalTeachers": { "type": "integer" },
                        "activeClasses": { "type": "integer" },
                        "performanceRating": { "type": "number" },
                        "contentHours": { "type": "integer" }
                      }
                    },
                    "applications": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "string" },
                          "name": { "type": "string" },
                          "title": { "type": "string" },
                          "timeText": { "type": "string" }
                        }
                      }
                    },
                    "topTeachers": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "string" },
                          "name": { "type": "string" },
                          "satisfactionRate": { "type": "integer" },
                          "avatar": { "type": "string" }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Bearer.\n\n- توكين STUDENT: يولَّد من `/auth/verify-otp` (صالح 30 يوم، يستخدم `JWT_SECRET`)\n- توكين ADMIN: يولَّد من `/auth/admin/login` (صالح 7 أيام، يستخدم `ADMIN_JWT_SECRET` أو `JWT_SECRET` كاحتياط)\n\n> ⚠️ **مهم**: لا تستخدم توكين STUDENT في endpoints الـ Admin والعكس صحيح — كلاهما سيرد بـ 401."
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Validation error - عادةً من class-validator على الحقول المطلوبة/الصيغة",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": {
              "statusCode": 400,
              "message": ["phone must match /^05\\d{8}$/ regular expression"],
              "error": "Bad Request"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "التوكن مفقود/منتهي/غير صالح أو الـ role لا يطابق الـ endpoint",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "examples": {
              "missingToken": { "summary": "بدون توكن", "value": { "statusCode": 401, "message": "Unauthorized" } },
              "badRole": { "summary": "توكين لا يطابق الـ role", "value": { "statusCode": 401, "message": "جلسة غير صالحة" } }
            }
          }
        }
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "statusCode": { "type": "integer", "example": 400 },
          "message": { "oneOf": [{ "type": "string" }, { "type": "array", "items": { "type": "string" } }] },
          "error": { "type": "string", "example": "Bad Request" }
        },
        "required": ["statusCode", "message"]
      },
      "DailyGoal": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "example": "goal_123" },
          "title": { "type": "string", "example": "مشاهدة فيديو الفيزياء (قوانين نيوتن)" },
          "completed": { "type": "boolean", "example": false },
          "dueDate": { "type": "string", "format": "date-time" }
        },
        "required": ["id", "title", "completed", "dueDate"]
      },
...
      "SubjectListItem": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string", "example": "الرياضيات" },
          "grade": { "type": "string", "enum": ["GRADE_11", "GRADE_12"] },
          "branch": { "type": "string", "enum": ["SCIENTIFIC", "LITERARY"] },
          "teacherName": { "type": "string", "nullable": true },
          "videoCount": { "type": "integer", "description": "عدد الفيديوهات المنشورة فقط" },
          "progressPercent": { "type": "integer", "description": "مستوى إنجاز المادة كنسبة مئوية" },
          "locked": { "type": "boolean", "description": "حالة قفل المادة بناء على الاشتراك" }
        }
      },
...
      "SubjectVideoItem": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "title": { "type": "string" },
          "description": { "type": "string", "nullable": true },
          "durationSec": { "type": "integer", "example": 645, "description": "مدة الفيديو بالثواني" },
          "unitNumber": { "type": "integer", "example": 1 },
          "orderInUnit": { "type": "integer", "example": 1 },
          "locked": { "type": "boolean", "description": "true إذا كان خارج حصة الخطة (بناءً على الترتيب)" }
        }
      },
      "SubjectVideosResponse": {
        "type": "object",
        "properties": {
          "subject": { "type": "object", "properties": { "id": { "type": "string" }, "name": { "type": "string" } } },
          "videos": { "type": "array", "items": { "$ref": "#/components/schemas/SubjectVideoItem" } },
          "quota": { "type": "integer", "description": "عدد الفيديوهات المتاحة من خطة الطالب" }
        }
      },
...
    }
  }
}